requirements.txt
aiofiles==23.2.1
aiogram==3.13.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
amocrm-api==2.6.1
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
coloredlogs==15.0.1
fasteners==0.19
frozenlist==1.4.1
humanfriendly==10.0
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
property-manager==3.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
verboselogs==1.7
yarl==1.11.1
